backend/requirements.txt
asgiref==3.8.1
attrs==25.1.0
boto3==1.37.20
botocore==1.37.20
colorama==0.4.6
Django==5.1.6
django-filter==25.1
django-redis==5.4.0
django-storages==1.14.5
djangorestframework==3.15.2
drf-spectacular==0.28.0
gunicorn==23.0.0
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
rpds-py==0.23.1
s3transfer==0.11.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
